Struct gearley::util::binary_heap::BinaryHeap

pub struct BinaryHeap<T> {
    // some fields omitted
}

A priority queue implemented with a binary heap.

This will be a max-heap.

Methods

impl<T: Ord> BinaryHeap<T>

fn new() -> BinaryHeap<T>

Creates an empty BinaryHeap as a max-heap.

fn with_capacity(capacity: usize) -> BinaryHeap<T>

Creates an empty BinaryHeap with a specific capacity. This preallocates enough memory for capacity elements, so that the BinaryHeap does not have to be reallocated until it contains at least that many values.

fn peek(&self) -> Option<&T>

Returns the greatest item in the binary heap, or None if it is empty.

fn capacity(&self) -> usize

Returns the number of elements the binary heap can hold without reallocating.

fn reserve_exact(&mut self, additional: usize)

Reserves the minimum capacity for exactly additional more elements to be inserted in the given BinaryHeap. Does nothing if the capacity is already sufficient.

fn reserve(&mut self, additional: usize)

Reserves capacity for at least additional more elements to be inserted in the BinaryHeap. The collection may reserve more space to avoid frequent reallocations.

fn shrink_to_fit(&mut self)

Discards as much additional capacity as possible.

fn pop(&mut self) -> Option<&T>

Removes the greatest item from the binary heap and returns it, or None if it is empty.

fn push(&mut self, item: T)

Pushes an item onto the binary heap.

fn len(&self) -> usize

Returns the length of the binary heap.

fn is_empty(&self) -> bool

Checks if the binary heap is empty.

fn clear(&mut self)

Drops all items from the binary heap.

fn as_data_slice(&self) -> &[T]

fn as_data_slice_mut(&mut self) -> &mut [T]